feat(player): implement movement with collision detection

+++ Common/src/main/java/model/entity/Bomb.java
@@ -5,7 +5,7 @@ public class Bomb extends Entity{
     private int radius;
     private int explosionDelay;
 
-    public Bomb(int ID, double x, double y, boolean isSolid, int ownerID, int radius, int explosionDelay) {
+    public Bomb(int ID, int x, int y, boolean isSolid, int ownerID, int radius, int explosionDelay) {
         super(ID, x, y, isSolid);
         this.ownerID = ownerID;
         this.radius = radius;
